+++ app/src/main/java/com/example/testappdl/MainActivity.kt
@@ -1,6 +1,6 @@
 package com.example.testappdl
 
-import android.annotation.SuppressLint
+
 import android.os.Bundle
 import android.util.Log
 import androidx.activity.ComponentActivity
