+++ app/src/main/java/com/example/testappdl/component/UserItem.kt
@@ -1,0 +1,30 @@
+package com.example.testappdl.component
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.example.testappdl.model.User.User
+
+@Composable
+fun UserItem(user: User, onClick: () -> Unit) {
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(8.dp)
+            .clickable { onClick.invoke() }
+    ) {
+        Column(modifier = Modifier.padding(16.dp)) {
+            Text(text = "Name:${user.name}")
+            Text(text = "Surname:${user.surname}")
+            Text(text = "Age ${user.age} years old")
+            Text(text = "Data destination: ${user.dataDestination.mark}")
+
+        }
+    }
+}
